% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invokeLambda.R
\name{rLambdaExecute}
\alias{rLambdaExecute}
\title{rLambdaExecute}
\usage{
rLambdaExecute()
}
\description{
Invoke a AWS Lambda function
}
\details{
This functon has no parameters. The entire local environment will be sent 
to the R server. The local environment should have a no-parameter function 
called \code{run}. This function will be run, and it's return value will be
placed in the local environment with name \code{result}. Additionally, 
any variables the \code{run()} function assigns to the enclosing 
environment will be present when \code{rLambdaExecute} returns.

(Note: this functionality is not working, for reasons still to be 
clarified. To get the final environment, it is necessary to 
execute \code{load("newdata.Rdata")}.)
}
