% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invokeLambda.R
\name{rLambdaInitialize}
\alias{rLambdaInitialize}
\title{rLambdaInitialize}
\usage{
rLambdaInitialize(path)
}
\arguments{
\item{path}{The path to the Python executable to be used.}
}
\description{
Initialize the interface to the Python 
engine and verify that required packages are installed.
}
\details{
By default the R package \code{reticulate} will check the 
system and use the first Python engine it finds. This is 
often not the one the user wants. A specific Python 
executable can be set by passing the path.

The interface to AWS is provided by the Python package
\code{boto3} and the request and response are sent via JSON.
The Python packages \code{boto3} and \code{json} must be installed
and available to the requested Python executable.
}
\examples{
rLambdaInitialize(path='/anaconda3/bin/python')

}
