% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invokeLambda.R
\name{rLambdaExecute}
\alias{rLambdaExecute}
\title{rLambdaExecute}
\usage{
rLambdaExecute()
}
\arguments{
\item{jsonin}{A JSON file with the input to the Lambda function}

\item{lambdafname}{The name of the Lambda function}
}
\value{
A list with the response key-value pairs.
}
\description{
Invoke a AWS Lambda function
}
\details{
The Lambda function must be already created. The user's 
default credentials will be used to certify the user's
permission to invoke the function. Input should be 
passed by means of a file with JSON content. This 
content will be sent to the Lambda function's 
\code{event} parameter.
}
